requirements.txt

# Core ML packages
torch==2.2.1
torchvision==0.17.1
numpy==1.25.2
scikit-learn==1.3.2
pillow==10.0.1
opencv-python==4.8.1.78

# Type checking
typeguard==4.5.1

# ONNX / inference
onnx==1.16.0
#onnxruntime==1.16.1
onnxruntime-gpu==1.16.1
onnxscript==0.4.0

# Utilities
PyYAML==6.0
jinja2==3.1.2

# API
fastapi==0.135.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6

# Monitoring
prometheus-client==0.18.0

# MLflow
mlflow==3.10.1
opentelemetry-proto==1.40.0



protobuf==5.29.3